requirements.txt
opencv-python
torch
torchvision
tqdm
pillow
numpy
